perfezionamento del codice

+++ src/main/java/com/example/springboot/services/impl/CarServiceImpl.java
@@ -73,6 +73,11 @@ public class CarServiceImpl implements CarService {
     public void delCar(long id) throws ItemNotFoundException {
         log.info("********** eliminazione auto : " + id + " **********");
         CarDTO car = getById(id);
+
+        if (car.getIdCar() == null) {
+            throw new ItemNotFoundException("auto non trovata");
+        }
+
         carRepository.delete(carMapper.newCar(car));
     }
 }
